refactor: UserDetails provider 리팩토링
