--- bbiyong-api/src/main/java/com/capstone/member/dto/MemberDto.java
@@ -1,62 +1,0 @@
-package com.capstone.member.dto;
-
-import com.capstone.member.domain.Member;
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-
-import java.time.LocalDateTime;
-import java.util.Collection;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class MemberDto implements UserDetails {
-
-    private Long id;
-    private String username;
-    private String password;
-    private LocalDateTime createdAt;
-    private LocalDateTime modifiedAt;
-
-    public static MemberDto of(Member member) {
-        return MemberDto.builder()
-                .id(member.getId())
-                .username(member.getUsername())
-                .password(member.getPassword())
-                .createdAt(member.getCreatedAt())
-                .modifiedAt(member.getModifiedAt())
-                .build();
-    }
-
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        return null;
-    }
-
-    @Override
-    public boolean isAccountNonExpired() {
-        return false;
-    }
-
-    @Override
-    public boolean isAccountNonLocked() {
-        return false;
-    }
-
-    @Override
-    public boolean isCredentialsNonExpired() {
-        return false;
-    }
-
-    @Override
-    public boolean isEnabled() {
-        return false;
-    }
-}
